# swagger.yaml
openapi: 3.0.3
info:
  title: Alliance Courtage API
  description: |
    API REST pour l'extranet Alliance Courtage.
    
    Cette API permet de gérer :
    - L'authentification et la gestion des utilisateurs
    - Les produits financiers et structurés
    - Les documents et archives
    - Les formations et notifications
    - Les simulateurs fiscaux
    - Le contenu CMS
    - Les réservations de produits
    
    **Authentification** : La plupart des endpoints nécessitent un token JWT dans le header `x-auth-token`.
  version: 1.0.0
  contact:
    name: Alliance Courtage
    email: support@alliance-courtage.fr
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement local
  - url: https://api.alliance-courtage.fr/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT dans le header x-auth-token
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID de l'utilisateur
        email:
          type: string
          format: email
          description: Email de l'utilisateur
        nom:
          type: string
          description: Nom de l'utilisateur
        prenom:
          type: string
          description: Prénom de l'utilisateur
        denomination_sociale:
          type: string
          nullable: true
          description: Dénomination sociale
        telephone:
          type: string
          nullable: true
          description: Téléphone
        code_postal:
          type: string
          nullable: true
          description: Code postal
        role:
          type: string
          enum: [admin, user, broker, client]
          description: Rôle de l'utilisateur
        is_active:
          type: boolean
          description: Statut actif/inactif
        created_at:
          type: string
          format: date-time
          description: Date de création
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Liste des utilisateurs
      description: Obtenir la liste de tous les utilisateurs (Admin uniquement)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, user, broker, client]
        - name: active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé (admin uniquement)
  /users/{id}:
    get:
      tags:
        - Users
      summary: Détails d'un utilisateur
      description: Obtenir les détails d'un utilisateur spécifique
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Utilisateur non trouvé

    put:
      tags:
        - Users
      summary: Modifier un utilisateur
      description: Modifier les informations d'un utilisateur (Admin uniquement)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                nom:
                  type: string
                prenom:
                  type: string
                denomination_sociale:
                  type: string
                  nullable: true
                telephone:
                  type: string
                  nullable: true
                code_postal:
                  type: string
                  nullable: true
                role:
                  type: string
                  enum: [admin, user, broker, client]
                is_active:
                  type: boolean
      responses:
        '200':
          description: Utilisateur modifié avec succès
        '403':
          description: Accès refusé (admin uniquement)

    delete:
      tags:
        - Users
      summary: Supprimer un utilisateur
      description: Supprimer un utilisateur (Admin uniquement)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Utilisateur supprimé avec succès
        '403':
          description: Accès refusé (admin uniquement)
  /users/{id}/change-password:
    put:
      tags:
        - Users
      summary: Changer le mot de passe
      description: Permettre à un utilisateur de changer son propre mot de passe
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Mot de passe changé avec succès
        '400':
          description: Mot de passe actuel incorrect
...